segmentation/requirements.txt
numpy==1.12.1
tensorflow==1.15.2
keras==2.1.4 
flask==1.0.2
pillow==6.2.0
h5py==2.8.0